Reworking the assembly AST to fix context assumptions
